chore: corrects visibility of DatabaseCombox
We now show DatabaseComboBox when even one of the identified dialects
requires database selection. To achieve this, instead of attaching the
first usable dialect, we now attach all the dialects that can be used
for the file, and for parsing we use the first of the identified
dialects. It is functionally still the same as what was happening
earlier, in addition, it helps us cover this conflicting case of
multiple detected dialects.