fix: identify command behind .iterator correctly
Earlier we were incorrectly identifying .iterator call always as a
FIND_MANY command. In reality .iterator could be chained on any
expression returning a MongoIterable.

This change, removes explicit identification of .iterator as FIND_MANY
command. IntelliJ eventually asks for a command for the right expression
which is generally the bottom most expression in the call chain.

The tests has been modified to replicate what IntelliJ does.