fix: skip inspection when command is not supported
A command is an identified mongodb command such as $match. There are a
few aggregate commands that our parsers do not support, in our case
anything other than match, group, unwind, sort, project, addFields, at
the time of this commit.

Earlier the parser was skipping the command that it does not support
but now we identify them and mark them as UNKNOWN so that later when we
are making decision wether to run inspection on the query or not, we can
decide that on the basis of whether the query has UNKNOWN command or
not.

The unknown operation detection logic has also been changed in this
commit. Earlier it was only looking in Filters component and instead it
should be looking at any component that has children for the Named
operations.