fix: correctly group inspection tools
We regressed in one of the earlier commits about how we group our
inspections in Settings -> Inspections panel in IntelliJ. This commits
fixes that.

The ideal grouping for our inspection is to be under one path which,
earlier, we were calling "MongoDB". That is one thing we changed in this
commit, populated groupPath metadata in plugin.xml for top level
grouping. Then we have subgrouped inspections in their categories using
groupKey metadata in plugin.xml.