chore: debounce CurrentQuery analyses on caret change
Earlier we were refreshing the scope analysis on every
keystroke -> caret changes and that was causing the
CurrentQuery scope to filter out the queries on EDT
thread which is expensive because in the filter method
we use CachedQueryService.queryAt which traverses the
Psi tree, and it comes at a cost.

Now we will debounce caret changes and refresh only after
a certain time as passed which will at-least keep us from
querying CachedQueryService to a minimum during active
development.

This is still far from ideal. Ideally we should assume
that operations such as tree traversal is costly and thus
should be done suspended but that is for another commit,
perhaps. This one should remove the immediate threat for
now.