fix: enforce the HasLimit req by Explain
ExplainQuery was adding HasLimit component with n=1 but when the query
already had a HasLimit component (parsed from the Java code), this newly
added component was getting ignored in the
DataGripMongoDbDriver.runQuery and instead of returning an object we
were returning a list, which was later causing an exception in
ExplainQuery itself.

I don't have a stack trace but it should be easy to reproduce by having
a query with limit and letting the Index inspection run on it. The
exception should be visible in debug console.

This commit fixes that by replacing the existing limit if any which
should be safe to do particularly for ExplainQuery.